requirements.txt
requests==2.31.0
Pillow==10.0.1
moviepy==1.0.3
gtts==2.3.2
pygame==2.5.2
imageio==2.31.5
python-dotenv==1.0.0